requirements.txt
# Core OCR framework
paddlepaddle>=2.6.0          # install paddlepaddle-gpu for GPU training
paddleocr>=2.7.3

# Numerics and data
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-image>=0.22.0         # skimage — required by PaddleOCR training (ppocr.data)

# PaddleOCR training (tools/train.py) — mirrors PaddleOCR/requirements.txt
albumentations>=1.3.0
albucore>=0.0.13
shapely>=2.0.0
pyclipper>=1.3.0
lmdb>=1.4.0
rapidfuzz>=3.0.0
opencv-python>=4.8.0
cython>=3.0.0
requests>=2.28.0

# Edit distance for CER/WER/DER
editdistance>=0.6.3

# Progress and logging
tqdm>=4.65.0
wandb>=0.16.0                # optional: set WANDB_MODE=disabled to skip

# Config serialisation
PyYAML>=6.0

# Visualisation (for analysis script)
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Baseline: Tesseract ──────────────────────────────────────────────────────
pytesseract>=0.3.10          # Python wrapper; also install: brew install tesseract tesseract-lang

# ── Baseline: Mistral OCR ────────────────────────────────────────────────────
mistralai>=1.0.0             # requires MISTRAL_API_KEY env var

# ── Baseline: Qwen 2.5 VL ───────────────────────────────────────────────────
transformers>=4.45.0
accelerate>=0.30.0
qwen-vl-utils>=0.0.8
# torch>=2.2.0               # install manually with correct CUDA version
# bitsandbytes>=0.43.0       # optional: 4-bit quantisation for Qwen on smaller GPUs
